Renames registry generator to azd extension
