+++ cli/azd/extensions/azd.internal.registry/internal/cmd/root.go
@@ -1,0 +1,27 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func NewRootCommand() *cobra.Command {
+	rootCmd := &cobra.Command{
+		Use:           "azd registry <command> [options]",
+		Short:         "Adds support to package AZD extensions for the registry.",
+		SilenceUsage:  true,
+		SilenceErrors: true,
+		CompletionOptions: cobra.CompletionOptions{
+			DisableDefaultCmd: true,
+		},
+	}
+
+	rootCmd.SetHelpCommand(&cobra.Command{Hidden: true})
+	rootCmd.PersistentFlags().Bool("debug", false, "Enable debug mode")
+
+	rootCmd.AddCommand(newPackageCommand())
+
+	return rootCmd
+}
