+++ cli/azd/extensions/azd.internal.registry/internal/helpers.go
